# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Doughmonster Worker API",
    "version": "1.0.0",
    "description": "OpenAPI metadata describing the public endpoints exposed by the Doughmonster Worker."
  },
  "servers": [
    {
      "url": "https://doughmonster-worker.thedoughmonster.workers.dev",
      "description": "Production deployment"
    },
    {
      "url": "http://127.0.0.1:8787",
      "description": "Local development (wrangler dev)"
    }
  ],
...
  "paths": {
    "/api/config/snapshot": {
      "get": {
        "summary": "Fetch Toast configuration snapshot",
        "description": "Fetches a fixed set of Toast configuration slices and caches the merged payload for one hour.",
        "tags": [
          "Configuration"
        ],
        "responses": {
          "200": {
            "description": "Configuration snapshot returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigSnapshotResponse"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/kitchen/prep-stations": {
      "get": {
        "summary": "List Toast kitchen prep stations",
        "description": "Returns the published Toast prep station configuration for the configured restaurant. Supports pagination via `pageToken` and incremental syncs with `lastModified`.",
        "tags": [
          "Kitchen"
        ],
        "responses": {
          "200": {
            "description": "Prep stations fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KitchenPrepStationsResponse"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/KitchenPrepStationsPageToken"
          },
          {
            "$ref": "#/components/parameters/KitchenPrepStationsLastModified"
          }
        ]
      }
    },
    "/api/menus": {
      "get": {
        "summary": "Retrieve the cached Toast menu document",
        "description": "Returns the currently cached Toast menus document along with cache metadata. Append `?refresh=1` (or any other truthy value) to force a synchronous refresh before responding.",
        "tags": [
          "Menus"
        ],
        "responses": {
          "200": {
            "description": "Menu document retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenusResponse"
                }
              }
            }
          },
          "502": {
            "description": "Upstream Toast API failure.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/MenusRefresh"
          }
        ]
      }
    },
    "/api/orders": {
      "get": {
        "summary": "Fetch the most recent Toast orders",
        "description": "Returns the most recent Toast orders using the worker's incremental KV-backed cache. By default the worker returns the five newest orders with full detail. Pass `detail=ids` to receive only the GUIDs while retaining ordering.",
        "tags": [
          "Orders"
        ],
        "responses": {
          "200": {
            "description": "Orders fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersLatestResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response emitted when the worker cannot return orders successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/OrdersLatestLimit"
          },
          {
            "$ref": "#/components/parameters/OrdersLatestDetail"
          },
          {
            "$ref": "#/components/parameters/OrdersLatestMinutes"
          },
          {
            "$ref": "#/components/parameters/OrdersLatestPageSize"
          },
          {
            "$ref": "#/components/parameters/OrdersLatestStart"
          },
          {
            "$ref": "#/components/parameters/OrdersLatestEnd"
          },
          {
            "$ref": "#/components/parameters/OrdersLatestSince"
          },
          {
            "$ref": "#/components/parameters/OrdersLatestBusinessDate"
          },
          {
            "$ref": "#/components/parameters/OrdersLatestTimeZone"
          },
          {
            "$ref": "#/components/parameters/OrdersLatestDebug"
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ConfigSnapshotResponse": {
        "type": "object",
        "properties": {
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "ttlSeconds": {
            "type": "integer"
          },
          "data": {
            "type": "object",
            "properties": {
              "diningOptions": {
                "oneOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "array",
                    "items": {}
                  },
                  {
                    "type": "string"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "Toast diningOptions slice payload when available."
              },
              "orderTypes": {
                "oneOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "array",
                    "items": {}
                  },
                  {
                    "type": "string"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "Toast orderTypes slice payload when available."
              },
              "revenueCenters": {
                "oneOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "array",
                    "items": {}
                  },
                  {
                    "type": "string"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "Toast revenueCenters slice payload when available."
              },
              "serviceAreas": {
                "oneOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "array",
                    "items": {}
                  },
                  {
                    "type": "string"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "Toast serviceAreas slice payload when available."
              },
              "taxRates": {
                "oneOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "array",
                    "items": {}
                  },
                  {
                    "type": "string"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "Toast taxRates slice payload when available."
              },
              "discounts": {
                "oneOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "array",
                    "items": {}
                  },
                  {
                    "type": "string"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "boolean"
                  },
                  {
                    "type": "null"
                  }
                ],
                "description": "Toast discounts slice payload when available."
              }
            },
            "required": [
              "diningOptions",
              "orderTypes",
              "revenueCenters",
              "serviceAreas",
              "taxRates",
              "discounts"
            ]
          }
        },
        "required": [
          "updatedAt",
          "ttlSeconds",
          "data"
        ],
        "additionalProperties": false
      },
...
      "ErrorObject": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "description": "Machine readable error code when available.",
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ErrorObject"
              },
              {
                "type": "string"
              }
            ]
          }
        },
        "required": [
          "ok",
          "error"
        ],
        "additionalProperties": true
      },
...
      "KitchenPrepStationsError": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": false
          },
          "route": {
            "type": "string",
            "const": "/api/kitchen/prep-stations"
          },
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "route",
          "error"
        ],
        "additionalProperties": true
      },
      "KitchenPrepStationsResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/KitchenPrepStationsSuccess"
          },
          {
            "$ref": "#/components/schemas/KitchenPrepStationsError"
          }
        ]
      },
      "KitchenPrepStationsSuccess": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "route": {
            "type": "string",
            "const": "/api/kitchen/prep-stations"
          },
          "count": {
            "type": "integer"
          },
          "prepStations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ToastPrepStation"
            }
          },
          "nextPageToken": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "request": {
            "type": "object",
            "properties": {
              "pageToken": {
                "oneOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "lastModified": {
                "oneOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "pageToken",
              "lastModified"
            ]
          },
          "raw": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "ok",
          "route",
          "count",
          "prepStations",
          "nextPageToken",
          "request"
        ]
      },
      "MenusResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/MenusSuccess"
          },
          {
            "$ref": "#/components/schemas/ErrorResponse"
          }
        ]
      },
      "MenusSuccess": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "menu": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ToastMenusDocument"
              },
              {
                "type": "null"
              }
            ]
          },
          "metadata": {
            "type": "object",
            "properties": {
              "lastUpdated": {
                "oneOf": [
                  {
                    "type": "string",
                    "format": "date-time"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "lastUpdated"
            ]
          },
          "cacheHit": {
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "menu",
          "metadata",
          "cacheHit"
        ]
      },
...
      "OrdersLatestBase": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "route": {
            "description": "Echoed route path.",
            "type": "string",
            "example": "/api/orders"
          },
          "limit": {
            "type": "integer"
          },
          "detail": {
            "type": "string",
            "enum": [
              "full",
              "ids"
            ]
          },
          "minutes": {
            "oneOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "window": {
            "type": "object",
            "properties": {
              "start": {
                "oneOf": [
                  {
                    "type": "string",
                    "format": "date-time"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "end": {
                "oneOf": [
                  {
                    "type": "string",
                    "format": "date-time"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "businessDate": {
                "type": "string"
              },
              "timeZone": {
                "type": "string"
              }
            },
            "required": [
              "start",
              "end"
            ]
          },
          "pageSize": {
            "type": "integer"
          },
          "expandUsed": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "count": {
            "type": "integer"
          },
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "orders": {
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ToastOrder"
                }
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "data": {
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ToastOrder"
                }
              },
              {
                "type": "null"
              }
            ]
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrdersLatestSource"
            }
          },
          "debug": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/OrdersLatestDebug"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "ok",
          "route",
          "limit",
          "detail",
          "minutes",
          "window",
          "pageSize",
          "expandUsed",
          "count",
          "ids"
        ]
      },
...
      "OrdersLatestError": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": false
          },
          "route": {
            "type": "string",
            "example": "/api/orders"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "route",
          "error"
        ],
        "additionalProperties": true
      },
      "OrdersLatestResponse": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/OrdersLatestSuccessFull"
          },
          {
            "$ref": "#/components/schemas/OrdersLatestSuccessIds"
          },
          {
            "$ref": "#/components/schemas/OrdersLatestError"
          }
        ]
      },
      "OrdersLatestSource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "enum": [
              "cache",
              "api",
              "merged"
            ]
          }
        },
        "required": [
          "id",
          "source"
        ]
      },
      "OrdersLatestSuccessFull": {
        "allOf": [
          {
            "$ref": "#/components/schemas/OrdersLatestBase"
          },
          {
            "type": "object",
            "properties": {
              "detail": {
                "type": "string",
                "enum": [
                  "full"
                ]
              },
              "orders": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ToastOrder"
                }
              },
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ToastOrder"
                }
              }
            },
            "required": [
              "orders"
            ]
          }
        ]
      },
      "OrdersLatestSuccessIds": {
        "allOf": [
          {
            "$ref": "#/components/schemas/OrdersLatestBase"
          },
          {
            "type": "object",
            "properties": {
              "detail": {
                "type": "string",
                "enum": [
                  "ids"
                ]
              },
              "orders": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "orders"
            ]
          }
        ]
      },
...
      "ToastMenusDocument": {
        "title": "Restaurant",
        "description": "Menus API restaurant payload containing published menus plus modifier references.",
        "type": "object",
        "properties": {
          "restaurantGuid": {
            "description": "Restaurant GUID associated with the published menus.",
            "format": "uuid",
            "type": "string"
          },
          "lastUpdated": {
            "description": "Timestamp when menu data was last published.",
            "type": "string"
          },
          "restaurantTimeZone": {
            "description": "Restaurant time zone expressed as an IANA identifier.",
            "type": "string"
          },
          "menus": {
            "description": "Published menus for the restaurant.",
            "type": "array",
            "items": {
              "description": "Menu",
              "type": "object"
            }
          },
          "modifierGroupReferences": {
            "description": "Dictionary of modifier groups keyed by referenceId.",
            "type": "object"
          },
          "modifierOptionReferences": {
            "description": "Dictionary of modifier options keyed by referenceId.",
            "type": "object"
          },
          "preModifierGroupReferences": {
            "description": "Dictionary of pre-modifier groups keyed by referenceId.",
            "type": "object"
          }
        }
      },
...
      "ToastPrepStation": {
        "title": "PrepStation",
        "description": "Kitchen prep station configuration representing a printer or KDS destination for routed orders.",
        "type": "object",
        "properties": {
          "guid": {
            "description": "Toast-maintained GUID for the referenced entity.",
            "type": "string"
          },
          "entityType": {
            "description": "Type discriminator returned by Toast (response only).",
            "type": "string"
          },
          "connectedPrepStations": {
            "description": "Other prep stations monitored by this station to coordinate multi-station orders.",
            "type": "array",
            "items": {
              "description": "Prep station reference",
              "type": "object",
              "properties": {
                "guid": {
                  "description": "Toast-maintained GUID for the referenced entity.",
                  "type": "string"
                },
                "entityType": {
                  "description": "Type discriminator returned by Toast (response only).",
                  "type": "string"
                }
              },
              "required": [
                "guid"
              ]
            }
          },
          "printingMode": {
            "description": "Printer routing behavior. `ON` prints every ticket, `OFFLINE_ONLY` only prints when the KDS is offline.",
            "enum": [
              "ON",
              "OFFLINE_ONLY"
            ],
            "type": "string"
          },
          "includeWithExpediter": {
            "description": "When true, also route tickets to the expediter device.",
            "type": "boolean"
          },
          "expoRouting": {
            "description": "Expo routing strategy controlling how tickets flow between the station and expediter.",
            "enum": [
              "SEND_TO_EXPO",
              "EXPO_ONLY",
              "SKIP_EXPO"
            ],
            "type": "string"
          },
          "name": {
            "description": "Human readable prep station name.",
            "type": "string"
          },
          "kitchenPrinter": {
            "description": "Printer associated with the prep station (if configured).",
            "type": "object",
            "properties": {
              "guid": {
                "description": "Toast-maintained GUID for the referenced entity.",
                "type": "string"
              },
              "entityType": {
                "description": "Type discriminator returned by Toast (response only).",
                "type": "string"
              }
            },
            "required": [
              "guid"
            ]
          }
        },
        "required": [
          "guid"
        ]
      },
...
    },
    "parameters": {
      "KitchenPrepStationsLastModified": {
        "name": "lastModified",
        "in": "query",
        "description": "Only return prep stations created or modified after the provided Toast timestamp (for example `2021-12-01T00:00:00.000+0000`).",
        "schema": {
          "type": "string"
        }
      },
      "KitchenPrepStationsPageToken": {
        "name": "pageToken",
        "in": "query",
        "description": "Pagination token returned by a previous Toast Kitchen API call. Forward the value to retrieve the next page.",
        "schema": {
          "type": "string"
        }
      },
      "MenusRefresh": {
        "name": "refresh",
        "in": "query",
        "description": "Force a synchronous refresh of the published menu when set to a truthy value.",
        "schema": {
          "type": "boolean"
        }
      },
...
      "OrdersLatestBusinessDate": {
        "name": "businessDate",
        "in": "query",
        "description": "Explicit Toast business date (yyyyMMdd) to anchor the fetch window.",
        "schema": {
          "type": "string",
          "pattern": "^\\d{8}$"
        }
      },
      "OrdersLatestDebug": {
        "name": "debug",
        "in": "query",
        "description": "Include extended diagnostics when `true` and the worker runs with `DEBUG` enabled.",
        "schema": {
          "type": "boolean"
        }
      },
      "OrdersLatestDetail": {
        "name": "detail",
        "in": "query",
        "description": "Toggle payload verbosity (`full` returns hydrated orders, `ids` returns GUIDs only).",
        "schema": {
          "type": "string",
          "enum": [
            "full",
            "ids"
          ],
          "default": "full"
        }
      },
      "OrdersLatestEnd": {
        "name": "end",
        "in": "query",
        "description": "ISO-8601 timestamp that constrains the upper bound of the fetch window.",
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "OrdersLatestLimit": {
        "name": "limit",
        "in": "query",
        "description": "Maximum number of orders to return (defaults to 5).",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 5
        }
      },
      "OrdersLatestMinutes": {
        "name": "minutes",
        "in": "query",
        "description": "Override the rolling window with an explicit lookback in minutes.",
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
      "OrdersLatestPageSize": {
        "name": "pageSize",
        "in": "query",
        "description": "Hint for the Toast bulk API page size (advanced diagnostics only).",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100
        }
      },
      "OrdersLatestSince": {
        "name": "since",
        "in": "query",
        "description": "Override the internal cursor with an ISO-8601 timestamp (debugging aid).",
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "OrdersLatestStart": {
        "name": "start",
        "in": "query",
        "description": "ISO-8601 timestamp that constrains the lower bound of the fetch window.",
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "OrdersLatestTimeZone": {
        "name": "timeZone",
        "in": "query",
        "description": "IANA time zone identifier used to interpret time-based parameters (defaults to UTC).",
        "schema": {
          "type": "string",
          "default": "UTC"
        }
      }
    }
  }
}